+++ changelog.md
@@ -15,3 +15,25 @@ fixed - main code
 added - function for mainChoice
 fixed - movement() so player can not walk off "map"
 created - a new replit for branches
+
+v2.0
+Name: RPG: Haunted Mansion - Inventory
+update - games name
+fixed - changelog name
+add - a dictionary to hold map-tile information (add description strings)
+update - change log
+fix - how room descriptions print in the game
+add - another dictionary for actions
+add - empty inventory, food, and medicine list
+add - a dictionary for randomizing results
+add - dictionary for all of the possible results
+add - a function for randomizing results
+update - mainChoice function. Now, walk is not the only option
+update - messages to be printed on the console
+
+v3.0
+Name: RPG: Haunted Mansion - Modules
+add - action file - a file for all of the action choices
+add - randomResults file - a file for randomizing results
+add - map file - a file for all of the map tiles and the direction function
+add - maptile text file - a file for the map